+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Support hexadecimal numbers parsing.
 - Initial Grok patterns support.
+- Log web server queries, update dependencies.
 
 ## [0.4.14] - 2023-11-15